added logins for google and facebook

+++ project.py
@@ -1,17 +1,41 @@
 from flask import Flask, render_template, request, redirect,jsonify, url_for, flash
+from flask import session as login_session
 app = Flask(__name__)
 
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Restaurant, MenuItem
+from database_setup import Base, Restaurant, MenuItem, User
+
+import random, string
+
+# creates a flow object from the client secrets JSON file that is storing the client ID
+# client secret, and other OAuth parameters
+from oauth2client.client import flow_from_clientsecrets
+
+# if we run into an error trying to exchange an authorization code for an access token
+# we can use the flow exchange error method to catch it.
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+
+# converts the returned value from a function into a real response that we
+# can send off to our client
+from flask import make_response
+import requests
+
+CLIENT_ID = json.loads(
+  open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "Restaurant Menu App Credentials"
 
 
 #Connect to Database and create database session
-engine = create_engine('sqlite:///restaurantmenu.db')
+engine = create_engine('sqlite:///restaurantmenuwithusers.db')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
+
+
 
 
 #JSON APIs to view Restaurant Information
@@ -43,8 +67,11 @@ def showRestaurants():
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
 def newRestaurant():
+  if 'username' not in login_session:
+    return redirct('/login')
   if request.method == 'POST':
-      newRestaurant = Restaurant(name = request.form['name'])
+      newRestaurant = Restaurant(name = request.form['name'],
+        user_id = login_session['user_id'])
       session.add(newRestaurant)
       flash('New Restaurant %s Successfully Created' % newRestaurant.name)
       session.commit()
@@ -55,6 +82,8 @@ def newRestaurant():
 #Edit a restaurant
 @app.route('/restaurant/<int:restaurant_id>/edit/', methods = ['GET', 'POST'])
 def editRestaurant(restaurant_id):
+  if 'username' not in login_session:
+    return redirect('/login')
   editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
       if request.form['name']:
@@ -68,7 +97,14 @@ def editRestaurant(restaurant_id):
 #Delete a restaurant
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods = ['GET','POST'])
 def deleteRestaurant(restaurant_id):
+  if 'username' not in login_session:
+    return redirect('/login')  
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  if restaurantToDelete.user_id != login_session['user_id']:
+    print "You are not authorized to delete."
+    return "<script>function myFunction() {alert('You are not authorized to delete \
+      this restaurant.  Please create your own restaurant in order to delete.');}</script> \
+      <body onload='myFunction()''>"
   if request.method == 'POST':
     session.delete(restaurantToDelete)
     flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -83,16 +119,27 @@ def deleteRestaurant(restaurant_id):
 def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-    return render_template('menu.html', items = items, restaurant = restaurant)
+    creator = getUserInfo(restaurant.user_id)
+    print creator.id
+    if 'user_id' in login_session and creator.id == login_session['user_id']:
+      print login_session['user_id']
+      print "showing private page"
+      return render_template('menu.html', items = items, restaurant = restaurant)
+    else:
+      print ("No user_id")
+      return render_template('publicmenu.html', items = items, restaurant = restaurant, creator = creator)
      
-
 
 #Create a new menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/new/',methods=['GET','POST'])
 def newMenuItem(restaurant_id):
+  if 'username' not in login_session:
+    return redirect('/login')
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
-      newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
+      newItem = MenuItem(name = request.form['name'], description = request.form['description'], 
+        price = request.form['price'], course = request.form['course'], 
+        restaurant_id = restaurant_id, user_id = restaurant.user_id)
       session.add(newItem)
       session.commit()
       flash('New Menu %s Item Successfully Created' % (newItem.name))
@@ -103,41 +150,295 @@ def newMenuItem(restaurant_id):
 #Edit a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET','POST'])
 def editMenuItem(restaurant_id, menu_id):
-
-    editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
-    restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-    if request.method == 'POST':
-        if request.form['name']:
-            editedItem.name = request.form['name']
-        if request.form['description']:
-            editedItem.description = request.form['description']
-        if request.form['price']:
-            editedItem.price = request.form['price']
-        if request.form['course']:
-            editedItem.course = request.form['course']
-        session.add(editedItem)
-        session.commit() 
-        flash('Menu Item Successfully Edited')
-        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
-    else:
-        return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
+  if 'username' not in login_session:
+    return redirect('/login')
+  editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
+  restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  if request.method == 'POST':
+      if request.form['name']:
+          editedItem.name = request.form['name']
+      if request.form['description']:
+          editedItem.description = request.form['description']
+      if request.form['price']:
+          editedItem.price = request.form['price']
+      if request.form['course']:
+          editedItem.course = request.form['course']
+      session.add(editedItem)
+      session.commit() 
+      flash('Menu Item Successfully Edited')
+      return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+  else:
+      return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
 
 
 #Delete a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete', methods = ['GET','POST'])
 def deleteMenuItem(restaurant_id,menu_id):
-    restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-    itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one() 
-    if request.method == 'POST':
-        session.delete(itemToDelete)
-        session.commit()
-        flash('Menu Item Successfully Deleted')
-        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
-    else:
-        return render_template('deleteMenuItem.html', item = itemToDelete)
+  if 'username' not in login_session:
+    return redirct('/login')
+  restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one() 
+  if request.method == 'POST':
+      session.delete(itemToDelete)
+      session.commit()
+      flash('Menu Item Successfully Deleted')
+      return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+  else:
+      return render_template('deleteMenuItem.html', item = itemToDelete)
+
+
+# Create a state token to prevent request forgery.
+# Store it in the session for later validation.
+@app.route('/login/')
+def showLogin():
+  state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x 
+    in xrange(32))
+  print state
+  login_session['state'] = state
+  return render_template('login.html', STATE = state)
+
+@app.route('/gconnect', methods = ['POST'])
+def gconnect():
+  print ("gconnect started.")
+  if request.args.get('state') != login_session['state']:
+    response = make_response(json.dumps('Invalid state parameter'), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  code = request.data
+
+  try: 
+    # Upgrade the authorization code into a credential object
+    oauth_flow = flow_from_clientsecrets('client_secrets.json', scope = '')
+    oauth_flow.redirect_uri = 'postmessage'
+    credentials = oauth_flow.step2_exchange(code)
+  except FlowExchangeError:
+    response = make_response(json.dumps("Failed to upgrade the authorization code."), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+
+  #Check that authentication token is valid.
+  access_token = credentials.access_token
+  url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token={}'.format(access_token))
+  h = httplib2.Http()
+  result = json.loads(h.request(url, 'GET')[1])
+  # If there was an error in the access token info, abort.
+  if result.get('error') is not None:
+    response = make_response(json.dumps(result.get('error')), 500)
+    reponse.headers['Content-Type'] = 'application/json'
+
+  # Verify that the access token is used for the intended user.
+  gplus_id = credentials.id_token['sub']
+  if result['user_id'] != gplus_id:
+    response = make_response(
+      json.dumps("Token's user ID doesn't match given user ID."), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+
+  # Verify that the access token is valid for this app.
+  if result['issued_to'] != CLIENT_ID:
+    response = make_response(
+      json.dumps("Token's client ID does not match app's."), 401)
+    print "Token's client ID does not match app's."
+    response.headers['Content-Type'] = 'application/json'
+    return response
+
+  # Check to see if user is already logged in.
+  stored_credentials = login_session.get('credentials')
+  stored_gplus_id = login_session.get('gplus_id')
+  if stored_credentials is not None and gplus_id == stored_gplus_id:
+    response = make_response(json.dumps('Current user already connected.'), 200)
+    response.headers['Content-Type'] = 'application/json'
+
+  # Store the access token in the session for later
+  login_session['provider'] = 'google'
+  login_session['credentials'] = credentials
+  login_session['gplus_id'] = gplus_id
+
+  # Get user info
+  userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+  params = {'access_token': credentials.access_token, 'alt': 'json'}
+  print params
+  answer = requests.get(userinfo_url, params = params)
+  data = answer.json()
+
+  login_session['username'] = data['name']
+  login_session['picture'] = data['picture']
+  login_session['email'] = data['email']
+
+  # See if user exists, if it doesn't make a new one.
+  user_id = getUserID(login_session['email'])
+  if not user_id:
+    user_id = createUser(login_session)
+  login_session['user_id'] = user_id
+
+  output = ''
+  output += '<h1>Welcome, '
+  output += login_session['username']
+  output += '!</h1>'
+  output += '<img src=" '
+  output += login_session['picture']
+  output += ' " style = "width: 300px; height: 300px; border-radius: 150px; '
+  output += ' -webkit-border-radius: 150px; -moz-border-radius: 150px;">'
+  flash('you are now logged in as {}'.format(login_session['username']))
+  return output
+  print "End of login code."
+
+@app.route('/fbconnect', methods = ['POST'])
+def fbconnect():
+  print ("fbconnect started.")
+  if request.args.get('state') != login_session['state']:
+    response = make_response(
+      json.dumps('Invalid state parameter.'), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+
+  access_token = request.data
+  print ("access token received {0}".format(access_token))
+
+  # Exchange client token for long-lived server-side token with GET /oauth/
+  # access_token?grant_type=fb_exchange_token&client_id={app-id}&client_sercret \
+  # ={app-secret}&fb_exchange_token={short-lived-token}
+  app_id = json.loads(open('fb_client_secrets.json', 'r').read())[
+      'web']['app_id']
+  app_secret = json.loads(
+      open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+  url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id={0}&client_secret={1}&fb_exchange_token={2}'.format(app_id, 
+    app_secret, access_token)
+  h = httplib2.Http()
+  result = h.request(url, 'GET')[1]
+
+  # Use token to get user info from API
+  userinfo_url = 'https://graph.facebook.com/v2.5/me'
+  # strip expire tag from access token
+  token = result.split("&")[0]
+
+  url = 'https://graph.facebook.com/v2.5/me?{0}&fields=name,id,email'.format(token)
+
+  h = httplib2.Http()
+  result = h.request(url, 'GET')[1]
+  # print ("url sent from API access:{0}".format(url))
+  # print "API JSON result: {0}".format(result)
+  data = json.loads(result)
+  login_session['provider'] = 'facebook'
+  login_session['username'] = data['name']
+  login_session['email'] = data['email']
+  login_session['facebook_id'] = data['id']
+
+  # The tokoen must be stored in the login_session in order to properly logout, let's strip out the information before the equals sign in our token.
+  stored_token = token.split("=")[1]
+  login_session['access_token'] = stored_token
+
+  # Get user picture
+  url = 'https://graph.facebook.com/v2.5/me/picture?{0}&redirect=0&height=200&width=200'.format(token)
+  h = httplib2.Http()
+  result = h.request(url, 'GET')[1]
+  data = json.loads(result)
+
+  login_session['picture'] = data['data']['url']
+
+  # See if user exists
+  user_id = getUserID(login_session['email'])
+  if not user_id:
+    user_id = createrUser(login_session)
+  login_session['user_id'] = user_id
+
+  output = ''
+  output += '<h1>Welcome, '
+  output += login_session['username']
+
+  output += '!</h1>'
+  output += '<img src='
+  output += login_session['picture']
+  output += ' "style = "width: 300px; height: 300px; border-radius: 150px; -webkit-border-radius: 150px; -moz-border-radius: 150px; -moz-border-radius: 150px;">'
+
+  flash("Now logged in as {0}".format(login_session['username']))
+  return output
+
+#DISCONNECT - Revoke a current user's token and reset their login_session.
+@app.route('/gdisconnect')
+def gdisconnect():
+  # Only disconnect a connecter user.
+  print login_session
+  credentials = login_session.get('credentials')
+  if credentials is None:  # if credentials is empty then nobody is logged in.
+    response = make_response(
+      json.dumps('Current user not connected.'), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+  # Execute HTTP GET request to revoke current token.
+  access_token = credentials.access_token
+  url = 'https://accounts.google.com/o/oauth2/revoke?token={0}'.format(access_token)
+  h = httplib2.Http()
+  result = h.request(url, "GET")[0]
+
+  if result['status'] != '200':
+    # For whatever reason, the given token was invalid.
+    response = make_response(
+      json.dumps('Failed to revoke token for given user.'), 400)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+
+@app.route('/fbdisconnect')
+def fbdisconnect():
+  facebook_id = login_session['facebook_id']
+  # The access token must be  included to successfully logout
+  access_token = login_session['access_token']
+  url = 'https://graph.facebook.com/{0}/permissions'.format(facebook_id)
+  h = httplib2.Http()
+  result = h.request(url, 'DELETE')[1]
+  return "you have been logged out"
+
+
+@app.route('/disconnect')
+def disconnect():
+  if 'provider' in login_session:
+    if login_session['provider'] == 'google':
+      gdisconnect()
+      del login_session['gplus_id']
+      del login_session['credentials']
+    if login_session['provider'] == 'facebook':
+      fbdisconnect()
+      del login_session['facebook_id']
+    del login_session['username']
+    del login_session['email']
+    del login_session['picture']
+    del login_session['user_id']
+    del login_session['provider']
+    flash ("You have been successfully logged out.")
+    return redirect(url_for('showRestaurants'))
+
+  else:
+    flash ("You were not logged in to begin with!")
+    return redirect(url_for('showRestaurants'))
+
+#####
 
 
 
+def getUserInfo(user_id):
+  user = session.query(User).filter_by(id = user_id).one()
+  return user
+
+def getUserID(email):
+  try:
+    user = session.query(User).filter_by(email = email).one()
+    return user.id
+  except:
+    return None
+
+def createUser(login_session):
+  newUser = User(name = login_session['username'], email = login_session['email'],
+    picture = login_session['picture'])
+  session.add(newUser)
+  session.commit()
+  user = session.query(User).filter_by(email = login_session['email']).one()
+  return user.id
+
+def getOwnerID(restaurant_id):
+  restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  ownerID = restaurant.user_id
+  # print ownerID
+  return ownerID
 
 if __name__ == '__main__':
   app.secret_key = 'super_secret_key'
